+++ chess4j-java/src/main/java/com/jamesswafford/chess4j/eval/EvalRook.java
@@ -55,7 +55,7 @@ public class EvalRook {
         return score;
     }
 
-    public static java.lang.Void extractRookFeatures(int[] features, Board board, Square sq, boolean endgame) {
+    public static java.lang.Void extractRookFeatures(double[] features, Board board, Square sq, boolean endgame) {
         boolean isWhite = board.getPiece(sq).isWhite();
         if (isWhite) {
             if (endgame) {
@@ -77,7 +77,7 @@ public class EvalRook {
         return null;
     }
 
-    private static void extractRookFeatures_OpenFile(int[] features, Board board, boolean isWhite, Square sq) {
+    private static void extractRookFeatures_OpenFile(double[] features, Board board, boolean isWhite, Square sq) {
         long friends,enemies;
         int inc;
         if (isWhite) {
